Implemented Dictionary CRUD actions
You can now Create, Read, Update and Delete (short CRUD, it's a thing,
look it up) Dictionaries. It also quite partly works with boards and
scores, but we'll have to sit together again and think about the Apps
architecture. From here on i will be quite quick to make adjustments to
the Database and the aviable methods to interact with it.

There's a DatabaseDemonstration Class that demonstrates how you
interact with the Database and dictionaries.

Don't hesitate to ask if you have a question!